% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkASRdqo.R
\name{checkASRdqo}
\alias{checkASRdqo}
\title{Check data quality objectives}
\usage{
checkASRdqo(dqodat)
}
\arguments{
\item{dqodat}{input data frame of data quality objectives}
}
\value{
\code{dqodat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check data quality objectives
}
\details{
This function is used internally within \code{\link{readASRdqo}} to run several checks on the input data to verify correct formatting before downstream analysis.

The following checks are made:
\itemize{
\item Column names: Should include only Parameter, Flag, GrMin, GrMax, Spike, FlatN, FlatDelta, RoCN, and RoCHours
\item All columns present: All columns from the previous check should be present
\item At least one parameter is present: At least one parameter in the \code{Parameter} column matches the \code{Parameter} column in \code{\link{paramsASR}}
\item Parameter format: All parameters listed in the \code{Parameter} column should match those in the \code{Parameter} column in \code{\link{paramsASR}}
\item Flag column: The \code{Flag} column should contain only "Fail" or "Suspect" entries
\item Rate of Change: No entries for Fail Flag rows
\item Numeric columns: All columns except \code{Parameter} and \code{Flag} should be numeric values
}
}
\examples{
library(dplyr)

dqopth <- system.file('extdata/ExampleDQO.xlsx', package = 'AquaSensR')

dqodat <- suppressWarnings(readxl::read_excel(dqopth, na = c('NA', 'na', ''),
     guess_max = Inf))

checkASRdqo(dqodat)
}
