% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkASRmeta.R
\name{checkASRmeta}
\alias{checkASRmeta}
\title{Check continuous monitoring metadata}
\usage{
checkASRmeta(metadat)
}
\arguments{
\item{metadat}{input data frame for continuous metadata}
}
\value{
\code{metadat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check continuous monitoring metadata
}
\details{
This function is used internally within \code{\link{readASRmeta}} to run several checks on the input data to verify correct formatting before downstream analysis.

The following checks are made:
\itemize{
\item Column names: Should include only Site, Parameter, Depth, Min, Max, Tlower, Tupper, SpikeFail, SpikeSuspect, FlatFailN, FlatFailDelta, FlatSuspectN, FlatSuspectDelta, RoCStdev, and RoCHours
\item All columns present: All columns from the previous check should be present
\item At least one parameter is present: At least one parameter in the \code{Parameter} column matches the \code{Parameter} column in \code{\link{paramsASR}}
\item Parameter format: All parameters listed in the \code{Parameter} column should match those in the \code{Parameter} column in \code{\link{paramsASR}}
\item Numeric columns: All columns except \code{Site} and \code{Parameter} should be numeric values
}
}
\examples{
library(dplyr)

metapth <- system.file('extdata/ExampleMeta.xlsx', package = 'AquaSensR')

metadat <- suppressWarnings(readxl::read_excel(metapth, na = c('NA', 'na', ''),
     guess_max = Inf))

checkASRmeta(metadat)
}
