% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formASRdqo.R
\name{formASRdqo}
\alias{formASRdqo}
\title{Format data quality objectives}
\usage{
formASRdqo(dqodat)
}
\arguments{
\item{dqodat}{input data frame}
}
\value{
A formatted data frame of the data quality objectives
}
\description{
Format data quality objectives
}
\details{
This function is used internally within \code{\link{readASRdqo}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item Convert non-numeric columns to numeric: Converts all columns except \code{Parameter} and \code{Flag} to numeric if they are not already.
}
}
\examples{
dqopth <- system.file('extdata/ExampleDQO.xlsx', package = 'AquaSensR')

dqodat <- suppressWarnings(readxl::read_excel(dqopth, na = c('NA', 'na', ''),
     guess_max = Inf))

formASRdqo(dqodat)
}
