% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formASRmeta.R
\name{formASRmeta}
\alias{formASRmeta}
\title{Format continuous metadata}
\usage{
formASRmeta(metadat)
}
\arguments{
\item{metadat}{input data frame}
}
\value{
A formatted data frame of the continuous data
}
\description{
Format continuous metadata
}
\details{
This function is used internally within \code{\link{readASRmeta}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item Convert non-numeric columns to numeric: Converts all columns except Site and Parameter to numeric if they are not already.
}
}
\examples{
metapth <- system.file('extdata/ExampleMeta.xlsx', package = 'AquaSensR')

metadat <- suppressWarnings(readxl::read_excel(metapth, na = c('NA', 'na', ''),
     guess_max = Inf))

formASRmeta(metadat)
}
