% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsASR.R
\docType{data}
\name{paramsASR}
\alias{paramsASR}
\title{Master list and units for acceptable parameters}
\format{
A \code{data.frame}
}
\usage{
paramsASR
}
\description{
Master list and units for acceptable parameters
}
\details{
This information is used to verify the correct format of input data and for formatting output data for upload to WQX.  A column showing the corresponding WQX names is also included.
}
\examples{
paramsASR
}
\keyword{datasets}
