% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASRmeta.R
\name{readASRmeta}
\alias{readASRmeta}
\title{Read continuous monitoring metadata from an external file}
\usage{
readASRmeta(metapth, runchk = TRUE)
}
\arguments{
\item{metapth}{character string of path to the metadata file}

\item{runchk}{logical to run data checks with \code{\link{checkASRmeta}}}
}
\value{
A formatted metadata data frame that can be used for downstream analysis
}
\description{
Read continuous monitoring metadata from an external file
}
\examples{

metapth <- system.file('extdata/ExampleMeta.xlsx', package = 'AquaSensR')

readASRmeta(metapth)
}
