% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkASRcont.R
\name{checkASRcont}
\alias{checkASRcont}
\title{Check continuous monitoring data}
\usage{
checkASRcont(contdat)
}
\arguments{
\item{contdat}{input data frame for results}
}
\value{
\code{contdat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check continuous monitoring data
}
\details{
This function is used internally within \code{\link{readASRcont}} to run several checks on the input data to verify correct formatting before downstream analysis.

The following checks are made:
\itemize{
\item: Column names: Should include only Site, Date, Time, and at least one parameter column that matches the \code{Label} column in \code{\link{paramsASR}}
\item: Site, Date, Time are present: These columns are required for downstream analysis and upload to WQX
\item: At least one parameter column is present: At least one parameter column that matches the \code{Label} column in \code{\link{paramsASR}} is required for downstream analysis and upload to WQX
\item Date format: Should be in a format that can be recognized by \code{\link[lubridate:ymd]{lubridate::ymd()}}
\item Time format: Should be in a format that can be recognized by \code{\link[lubridate:ymd_hms]{lubridate::ymd_hms()}}
\item Parameter columns should be numeric: All parameter columns should be numeric values
}
}
\examples{
library(dplyr)

contpth <- system.file('extdata/ExampleCont.xlsx', package = 'AquaSensR')

contdat <- suppressWarnings(readxl::read_excel(contpth, na = c('NA', 'na', ''),
     guess_max = Inf)) |>
   dplyr::mutate(dplyr::across(
     dplyr::where(~ inherits(.x, "POSIXct") | inherits(.x, "Date")),
   as.character))

checkASRcont(contdat)
}
