% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formASRcont.R
\name{formASRcont}
\alias{formASRcont}
\title{Format continuous data}
\usage{
formASRcont(contdat, tz)
}
\arguments{
\item{contdat}{input data frame}

\item{tz}{character string of time zone for the date and time columns  See \code{OlsonNames()} for acceptable time zones.}
}
\value{
A formatted data frame of the continuous data
}
\description{
Format continuous data
}
\details{
This function is used internally within \code{\link{readASRcont}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item Combine Date and Time columns: Combines into a single DateTime column, converts to POSIXct with the specified time zone.
\item Convert non-numeric columns to numeric: Converts all columns except Site and DateTime to numeric if they are not already.
}
}
\examples{
contpth <- system.file('extdata/ExampleCont.xlsx', package = 'AquaSensR')

contdat <- suppressWarnings(readxl::read_excel(contpth, na = c('NA', 'na', ''),
     guess_max = Inf)) |>
   dplyr::mutate(dplyr::across(
     dplyr::where(~ inherits(.x, "POSIXct") | inherits(.x, "Date")),
   as.character))

formASRcont(contdat, tz = 'Etc/GMT+5')
}
