% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASRcont.R
\name{readASRcont}
\alias{readASRcont}
\title{Read continuous monitoring data from an external file}
\usage{
readASRcont(contpth, tz = "Etc/GMT+5", runchk = TRUE)
}
\arguments{
\item{contpth}{character string of path to the continuous data file}

\item{tz}{character string of time zone for the date and time columns, defaults to Etc/GMT+5 (Eastern time zone, no daylight savings).  See \code{OlsonNames()} for acceptable time zones.}

\item{runchk}{logical to run data checks with \code{\link{checkASRcont}}}
}
\value{
A formatted continuous monitoring data frame that can be used for downstream analysis
}
\description{
Read continuous monitoring data from an external file
}
\details{
The file is imported via \code{\link{utilASRimportcont}}, which forces
\code{Date}, \code{Time}, and \code{DateTime} columns to character and
converts Excel numeric serial representations to human-readable strings
before checks are run.

Always verify the correct time zone for your data.  If your data are in a different time zone than Etc/GMT+5 (default), specify the correct time zone in the \code{tz} argument.
}
\examples{

contpth <- system.file('extdata/ExampleCont2.xlsx', package = 'AquaSensR')

readASRcont(contpth)
}
