% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASRcont.R
\name{readASRcont}
\alias{readASRcont}
\title{Read continuous monitoring data from an external file}
\usage{
readASRcont(contpth, tz, runchk = TRUE)
}
\arguments{
\item{contpth}{character string of path to the continuous data file}

\item{tz}{character string of time zone for the date and time columns  See \code{OlsonNames()} for acceptable time zones.}

\item{runchk}{logical to run data checks with \code{\link{checkASRcont}}}
}
\value{
A formatted continuous monitoring data frame that can be used for downstream analysis
}
\description{
Read continuous monitoring data from an external file
}
\examples{

contpth <- system.file('extdata/ExampleCont.xlsx', package = 'AquaSensR')

readASRcont(contpth, tz = 'Etc/GMT+5')
}
