% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASRcont.R
\name{readASRcont}
\alias{readASRcont}
\title{Read continuous monitoring data from an external file}
\usage{
readASRcont(contpth, runchk = T)
}
\arguments{
\item{contpth}{character string of path to the results file}

\item{runchk}{logical to run data checks with \code{\link{checkASRcont}}}
}
\value{
A formatted continuous monitoring data frame that can be used for downstream analysis
}
\description{
Read continuous monitoring data from an external file
}
\examples{

contpth <- system.file('extdata/ExampleCont.xlsx', package = 'AquaSensR')

readASRcont(contpth)
}
