% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASRdqo.R
\name{readASRdqo}
\alias{readASRdqo}
\title{Read data quality objectives from an external file}
\usage{
readASRdqo(dqopth, runchk = TRUE)
}
\arguments{
\item{dqopth}{character string of path to the data quality objectives file}

\item{runchk}{logical to run data checks with \code{\link{checkASRdqo}}}
}
\value{
A formatted data quality objectives data frame that can be used for downstream analysis
}
\description{
Read data quality objectives from an external file
}
\examples{

dqopth <- system.file('extdata/ExampleDQO.xlsx', package = 'AquaSensR')

readASRdqo(dqopth)
}
